@@ -14,83 +14,83 @@ use ZF\OAuth2\Adapter\PdoAdapter as ZFOAuthPdoAdapter;
  */
 class PdoAdapter extends ZFOAuthPdoAdapter
 {
-	/**
+    /**
      * @param string $connection
-     * @param array $config
+     * @param array  $config
      */
-    public function __construct($connection, $config = array())
+    public function __construct($connection, $config = [])
     {
-    	$config = array(
+        $config = [
             'user_provider_table' => 'oauth_user_provider',
             'user_provider_access_token_table' => 'oauth_user_provider_access_tokens'
-        );
-        parent::__construct($connection, $config);        
+        ];
+        parent::__construct($connection, $config);
     }
 
     public function setUserProvider($provider, $provider_id, $username)
     {
-    	// if it exists, update it.
-    	if($this->getUserProvider($username, $provider)) {
-    		$stmt = $this->db->prepare(sprintf(
+        // if it exists, update it.
+        if ($this->getUserProvider($username, $provider)) {
+            $stmt = $this->db->prepare(sprintf(
                 'UPDATE %s SET provider_id=:provider_id where user_id=:username and provider=:provider',
                 $this->config['user_provider_table']
             ));
-    	} else {
-    		$stmt = $this->db->prepare(sprintf(
+        } else {
+            $stmt = $this->db->prepare(sprintf(
                 'INSERT INTO %s (provider, provider_id, user_id) VALUES (:provider, :provider_id, :username)',
                 $this->config['user_provider_table']
             ));
-    	}
+        }
 
-    	return $stmt->execute(compact('provider', 'provider_id', 'username'));
+        return $stmt->execute(compact('provider', 'provider_id', 'username'));
     }
 
     public function getUserProvider($username, $provider)
     {
         $stmt = $this->db->prepare($sql = sprintf('SELECT * from %s where user_id=:username and provider=:provider', $this->config['user_provider_table']));
-        $stmt->execute(array(
-        	'username' => $username,
-        	'provider' => $provider
-        ));
+        $stmt->execute([
+            'username' => $username,
+            'provider' => $provider
+        ]);
 
         if (!$providerInfo = $stmt->fetch()) {
             return false;
         }
 
         // the default behavior is to use "username" as the user_id
-        return array_merge(array(
+        return array_merge([
             'user_id' => $username
-        ), $providerInfo);
+        ], $providerInfo);
     }
 
     public function setUserProviderAccessToken($access_token, $provider, $provider_id, $user_id, $expires = null, $scope = null)
     {
-    	// convert expires to datestring
+        // convert expires to datestring
         $expires = date('Y-m-d H:i:s', $expires);
 
-    	// if it exists, update it.
-    	if($this->getUserProviderAccessToken($access_token, $provider, $provider_id, $user_id)) {
-    		$stmt = $this->db->prepare(sprintf(
+        // if it exists, update it.
+        if ($this->getUserProviderAccessToken($access_token, $provider, $provider_id, $user_id)) {
+            $stmt = $this->db->prepare(sprintf(
                 'UPDATE %s SET provider=:provider, provider_id=:provider_id, user_id=:user_id, expires=:expires, scope=:scope where access_token=:access_token',
                 $this->config['user_provider_access_token_table']
             ));
-    	} else {
-    		$stmt = $this->db->prepare(sprintf(
+        } else {
+            $stmt = $this->db->prepare(sprintf(
                 'INSERT INTO %s (access_token, provider, provider_id, user_id, expires, scope) VALUES (:access_token, :provider, :provider_id, :user_id, :expires, :scope)',
                 $this->config['user_provider_access_token_table']
             ));
-    	}
+        }
 
-    	return $stmt->execute(compact('access_token', 'provider', 'provider_id', 'user_id', 'expires', 'scope'));
+        return $stmt->execute(compact('access_token', 'provider', 'provider_id', 'user_id', 'expires', 'scope'));
     }
 
     public function getUserProviderAccessToken($access_token, $provider, $provider_id, $user_id)
     {
         $stmt = $this->db->prepare(
-        	$sql = sprintf('SELECT * from %s where access_token=:access_token and provider=:provider and provider_id=:provider_id and user_id=:user_id', 
-        	$this->config['user_provider_access_token_table']
+            $sql = sprintf('SELECT * from %s where access_token=:access_token and provider=:provider and provider_id=:provider_id and user_id=:user_id',
+            $this->config['user_provider_access_token_table']
         ));
-        
+
         $token = $stmt->execute(compact('access_token', 'provider', 'provider_id', 'user_id'));
 
         if ($token = $stmt->fetch()) {
@@ -101,14 +101,13 @@ class PdoAdapter extends ZFOAuthPdoAdapter
         return $token;
     }
 
-
     /**
      * Set the user
      *
-     * @param string $username
-     * @param string $password
-     * @param string $firstName
-     * @param string $lastName
+     * @param  string $username
+     * @param  string $password
+     * @param  string $firstName
+     * @param  string $lastName
      * @return bool
      */
     public function setUser($username, $password, $firstName = null, $lastName = null)
