@@ -12,6 +12,6 @@ class Odnoklassniki extends \Hybrid_Providers_Odnoklassniki
     public $curl_auth_header      = true;
     public $curl_useragent        = "OAuth/2 Simple PHP Client v0.1; HybridAuth http://hybridauth.sourceforge.net/";
     public $curl_proxy            = null;
-    public $curl_header           = array("Accept: application/json; charset=UTF-8");
+    public $curl_header           = ["Accept: application/json; charset=UTF-8"];
     public $http_info             = "";
 }
